현재 사용자의 인증 정보 받아오기 및 검증

+++ src/main/java/com/example/usedmarket/service/NegoService.java
@@ -4,13 +4,16 @@ import com.example.usedmarket.dto.NegoDto;
 import com.example.usedmarket.dto.SalesItemDto;
 import com.example.usedmarket.entity.NegoEntity;
 import com.example.usedmarket.entity.SalesItemEntity;
+import com.example.usedmarket.entity.UserEntity;
 import com.example.usedmarket.repo.NegoRepo;
 import com.example.usedmarket.repo.SalesItemRepo;
+import com.example.usedmarket.repo.UserRepo;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.http.HttpStatus;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 import org.springframework.web.server.ResponseStatusException;
 
@@ -24,17 +27,23 @@ import java.util.Optional;
 public class NegoService {
     private final NegoRepo negoRepository;
     private final SalesItemRepo itemRepository;
+    private final UserRepo userRepository;
 
     // 제안 등록
     public NegoDto createNego(Long itemId, NegoDto dto) {
+        // 현재 사용자의 인증 정보 가져오기
+        String auth = SecurityContextHolder.getContext().getAuthentication().getName();
+        UserEntity user = userRepository.findAllByUsername(auth);
+
         SalesItemEntity itemEntity = itemCheck(itemId);
+
         NegoEntity negoEntity = new NegoEntity();
         negoEntity.setItemId(itemId);
-        negoEntity.setSalesItem(itemEntity);
+//        negoEntity.setSalesItem(itemEntity);
+        negoEntity.setAddItem(itemEntity);
         negoEntity.setSuggestedPrice(dto.getSuggestedPrice());
         negoEntity.setStatus(dto.getStatus());
-        negoEntity.setWriter(dto.getWriter());
-        negoEntity.setPassword(dto.getPassword());
+        negoEntity.setAddUser(user);
         return NegoDto.fromEntity(negoRepository.save(negoEntity));
     }
 
@@ -50,31 +59,58 @@ public class NegoService {
     }
 
     public NegoDto updateNego(Long itemId, Long id, NegoDto dto) throws IllegalAccessException {
-        // 게시글이 존재하는 지 확인
+        // 현재 사용자의 인증 정보 가져오기
+        String auth = SecurityContextHolder.getContext().getAuthentication().getName();
+        UserEntity user = userRepository.findAllByUsername(auth);
+
+        // 게시글, 제안 id 존재하는 지 확인
         SalesItemEntity itemEntity = itemCheck(itemId);
-        // 제안 id 존재 및 유저 검증
-        NegoEntity negoEntity = userCheck(id, dto.getWriter(), dto.getPassword());
+        NegoEntity negoEntity = negoCheck(id);
+
+        // 물품 등록시 사용했던 계정과 일치하다면
+        if (!user.equals(negoEntity.getAddUser()))
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST);
+
         // 다른 문제 없을 시 내용 수정
         negoEntity.setSuggestedPrice(dto.getSuggestedPrice());
-        negoEntity.setStatus(dto.getStatus());
         // 수정된 내용 entity에 저장
         return NegoDto.fromEntity(negoRepository.save(negoEntity));
     }
 
-    public void deleteNego(Long itemId, Long id, NegoDto dto) throws IllegalAccessException {
+    public void deleteNego(Long itemId, Long id) throws IllegalAccessException {
+        // 현재 사용자의 인증 정보 가져오기
+        String auth = SecurityContextHolder.getContext().getAuthentication().getName();
+        UserEntity user = userRepository.findAllByUsername(auth);
+
         SalesItemEntity itemEntity = itemCheck(itemId);
-        NegoEntity negoEntity = userCheck(id, dto.getWriter(), dto.getPassword());
+        NegoEntity negoEntity = negoCheck(id);
+
+        // 물품 등록시 사용했던 계정과 일치하다면
+        if (!user.equals(negoEntity.getAddUser()))
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST);
+
         negoRepository.deleteById(id);
     }
 
 
-    // 대상 물품 주인이 status 변경 (수락 및 거절)
-    // 물품 등록 시 사용했던 작성자와 비밀번호 첨부
+    /*
+    대상 물품의 주인은 구매 제안을 수락 | 거절 할 수 있다.
+    1. 이를 위해서 제안의 대상 물품을 등록할 때 사용한 **작성자와 비밀번호**를 첨부해야 한다.
+    2. 이때 구매 제안의 상태는 **수락** | **거절** 이 된다.
+    */
     public NegoDto acceptOrReject(Long itemId, Long id, NegoDto dto) throws IllegalAccessException {
-        // 제안 대상 물품 찾기 및 등록자 정보 검증
-        SalesItemEntity itemEntity = itemUserCheck(itemId, dto.getWriter(), dto.getPassword());
-        // 제안 id 찾기
+        // 현재 사용자의 인증 정보 가져오기
+        String auth = SecurityContextHolder.getContext().getAuthentication().getName();
+        UserEntity user = userRepository.findAllByUsername(auth);
+
+        // 제안 대상 물품, 제안 id 찾기
+        SalesItemEntity itemEntity = itemCheck(itemId);
         NegoEntity negoEntity = negoCheck(id);
+
+        // 물품을 등록한 사용자 계정인 것이 인증되면 수락 | 거절 가능
+        if(!user.equals(itemEntity.getAddUser()))
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST);
+
         // 물품 등록자는 구매 제안 상태 변경 가능 (수락 | 거절)
         negoEntity.setStatus(dto.getStatus());
         return NegoDto.fromEntity(negoRepository.save(negoEntity));
@@ -85,8 +121,17 @@ public class NegoService {
     // 제안의 상태는(status) '확정' 으로 변경 -> '판매완료' 상태가 된다.
     // 구매 제안이 확정될 경우, 확정되지 않은 다른 구매 제의 상태는 모두 거절이 된다.
     public NegoDto purchaseConfirm(Long itemId, Long id, NegoDto dto) throws IllegalAccessException {
+        // 현재 사용자의 인증 정보 가져오기
+        String auth = SecurityContextHolder.getContext().getAuthentication().getName();
+        UserEntity user = userRepository.findAllByUsername(auth);
+
         SalesItemEntity itemEntity = itemCheck(itemId);
-        NegoEntity negoEntity = userCheck(id, dto.getWriter(), dto.getPassword());
+        NegoEntity negoEntity = negoCheck(id);
+
+        // 제안 등록 시 사용했던 작성자와 비밀번호가 일치한다면 다음 로직 수행
+        if(!user.equals(negoEntity.getAddUser()))
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST);
+
         // 수락일 때 : 제안 등록자는 '확정'을 할 수 있다.
         if(negoEntity.getStatus().equals("수락"))
             negoEntity.setStatus(dto.getStatus());
@@ -97,7 +142,7 @@ public class NegoService {
             negoRepository.save(negoEntity);
             // 구매 확정 시 다른 제안 거절
             if(negoEntity.getStatus().equals("판매 완료")){
-                List<NegoEntity> negotiationEntities = negoRepository.findBySalesItemId(itemId);
+                List<NegoEntity> negotiationEntities = negoRepository.findByItemId(itemId);
                 for (NegoEntity entity : negotiationEntities) {
                     if (!entity.getId().equals(id)) {
                         entity.setStatus("거절");
@@ -126,26 +171,26 @@ public class NegoService {
     }
 
     // 제안 유저 검증
-    public NegoEntity userCheck(Long id, String writer, String password ) throws IllegalAccessException {
-        Optional<NegoEntity> optionalNego = negoRepository.findById(id);
-        if(optionalNego.isEmpty())
-            throw new ResponseStatusException(HttpStatus.NOT_FOUND);
-
-        NegoEntity entity = optionalNego.get();
-        if(!entity.getWriter().equals(writer) || !entity.getPassword().equals(password))
-            throw new IllegalAccessException("사용자 정보가 일치하지 않습니다.");
-
-        return entity;
-    }
+//    public NegoEntity userCheck(Long id, String writer, String password ) throws IllegalAccessException {
+//        Optional<NegoEntity> optionalNego = negoRepository.findById(id);
+//        if(optionalNego.isEmpty())
+//            throw new ResponseStatusException(HttpStatus.NOT_FOUND);
+//
+//        NegoEntity entity = optionalNego.get();
+//        if(!entity.getWriter().equals(writer) || !entity.getPassword().equals(password))
+//            throw new IllegalAccessException("사용자 정보가 일치하지 않습니다.");
+//
+//        return entity;
+//    }
 
     // 물품 유저 검증
-    public SalesItemEntity itemUserCheck(Long itemId, String writer, String password) throws IllegalAccessException {
-        Optional<SalesItemEntity> optionalItem = itemRepository.findById(itemId);
-        if(optionalItem.isEmpty())
-            throw new ResponseStatusException(HttpStatus.NOT_FOUND);
-        SalesItemEntity entity = optionalItem.get();
-        if(!entity.getWriter().equals(writer) || !entity.getPassword().equals(password))
-            throw new IllegalAccessException("물품 등록자 정보와 일치하지 않습니다.");
-        return entity;
-    }
+//    public SalesItemEntity itemUserCheck(Long itemId, String writer, String password) throws IllegalAccessException {
+//        Optional<SalesItemEntity> optionalItem = itemRepository.findById(itemId);
+//        if(optionalItem.isEmpty())
+//            throw new ResponseStatusException(HttpStatus.NOT_FOUND);
+//        SalesItemEntity entity = optionalItem.get();
+//        if(!entity.getWriter().equals(writer) || !entity.getPassword().equals(password))
+//            throw new IllegalAccessException("물품 등록자 정보와 일치하지 않습니다.");
+//        return entity;
+//    }
 }
